# hackathon-backend/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Hackathon, VMO-Global Car Sharing
  description: 'VMO-Global Car Sharing Api Docs'
  version: 1.0.0
servers:
- url: http://localhost:8082/api/v1/
...
paths:
...
  /car/{id}:
    get:
      tags:
      - Car
      summary: Find details car by ID
      description: Returns details car
      operationId: getCarById
      parameters:
      - name: id
        in: path
        description: ID of car to return
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                  data: 
                    type: object
                    properties: 
                    $ref: '#/components/schemas/Car'
        500:
          description: Internal Server Error
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "CONTRACT_CONTEXTUAL_ERROR"
                  message: 
                    type: string
                    description: "Car does not exist!"
    put:
      tags:
      - Car
      summary: Updates a car in the store with form data
      operationId: updateCar
      parameters:
      - name: id
        in: path
        description: ID of car that needs to be updated
        required: true
        schema:
          type: string
      requestBody:
        content:
            application/json:
              schema:
                type: object
                properties: 
                  categoryCode: 
                    type: string
                  name: 
                    type: string
                  description: 
                    type: string
                  imageUrl: 
                    type: string
                  pricePerDay: 
                    type: number
                    format: double
                    default: 0
                  updateBy: 
                    type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "Succeeded!"
                  message: 
                    type: string
                    description: "Car update successfully!"
        500:
          description: Internal Server Error
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "CONTRACT_CONTEXTUAL_ERROR"
                  message: 
                    type: string
                    description: "Car does not exist!"
  
    delete:
      tags:
      - Car
      summary: Delete a car in the store with ID
      operationId: deleteCar
      parameters:
      - name: id
        in: path
        description: ID of car that needs to be updated
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "Succeeded!"
                  message: 
                    type: string
                    description: "Car deleted successfully!"
        500:
          description: Internal Server Error
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "CONTRACT_CONTEXTUAL_ERROR"
                  message: 
                    type: string
                    description: "Car does not exist!"
...
  /car-rental:
    get:
      tags:
      - Rental
      summary: List of rentals
      description: list rentals
      operationId: geListRentals
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                  data: 
                    type: array
                    items:
                      $ref: '#/components/schemas/Rental'
        500:
          description: "Internal Network Error"
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "UNKNOWN_TRANSACTION_STATUS"
                  message: 
                    type: string
                    description: "UNAVAILABLE: io exception"
    post:
      tags:
      - Rental
      summary: Add a new rental
      operationId: addrental
      requestBody:
        description: Rental object that needs to be added
        content:
          application/json:
            schema:
              type: object
              properties: 
                carId: 
                  type: string
                customerId: 
                  type: string
                fromDate: 
                  type: string
                  description: "2021-03-18 11:12:33+0700"
                dueDate: 
                  type: string
                  description: "2021-03-19 11:12:33+0700"
                createBy: 
                  type: string
        required: true
      responses:
        200:
          description: "Successfully"
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                  message: 
                    type: string
        500:
          description: "Internal Network Error"
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "UNKNOWN_TRANSACTION_STATUS"
                  message: 
                    type: string
                    description: "UNAVAILABLE: io exception"
  /car-rental/{id}:
    get:
      tags:
      - Rental
      summary: Find details rental by ID
      description: Returns details rental
      operationId: getRentalById
      parameters:
      - name: id
        in: path
        description: ID of rental to return
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                  data: 
                    $ref: '#/components/schemas/Rental'
        500:
          description: Internal Server Error
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "CONTRACT_CONTEXTUAL_ERROR"
                  message: 
                    type: string
                    description: "Car Rental does not exist!"
    put:
      tags:
      - Rental
      summary: Updates a rental in the store with form data
      operationId: updateRental
      parameters:
      - name: id
        in: path
        description: ID of rental that needs to be updated
        required: true
        schema:
          type: string
      requestBody:
        content:
            application/json:
              schema:
                type: object
                properties: 
                  carId: 
                    type: string
                  customerId: 
                    type: string
                  fromDate: 
                    type: string
                    description: "2021-03-10 11:12:33+0700"
                  dueDate: 
                    type: string
                    description: "2021-03-12 11:12:33+0700"
                  updateBy: 
                    type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "Succeeded!"
                  message: 
                    type: string
                    description: "Car Rental update successfully!"
        500:
          description: Internal Server Error
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "CONTRACT_CONTEXTUAL_ERROR"
                  message: 
                    type: string
                    description: "Car Rental does not exist!"
  /car-rental/{id}/return:
    put:
      tags:
      - Rental
      summary: Return a rental
      operationId: returnRental
      parameters:
      - name: id
        in: path
        description: ID of rental that needs to be return
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: 
                updateBy: 
                  type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "Succeeded!"
                  message: 
                    type: string
                    description: "Rental update successfully!"
        500:
          description: Internal Server Error
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "CONTRACT_CONTEXTUAL_ERROR"
                  message: 
                    type: string
                    description: "Rental does not exist!"
  /car-rental/{id}/start:
    put:
      tags:
      - Rental
      summary: Start a rental
      operationId: startRental
      parameters:
      - name: id
        in: path
        description: ID of rental that needs to be start
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: 
                updateBy: 
                  type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "Succeeded!"
                  message: 
                    type: string
                    description: "Rental update successfully!"
        500:
          description: Internal Server Error
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "CONTRACT_CONTEXTUAL_ERROR"
                  message: 
                    type: string
                    description: "Rental does not exist!"
  /car-rental/{id}/cancel:
    put:
      tags:
      - Rental
      summary: Cancel a rental
      operationId: cancelRental
      parameters:
      - name: id
        in: path
        description: ID of rental that needs to be cancel
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: 
                updateBy: 
                  type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "Succeeded!"
                  message: 
                    type: string
                    description: "Rental update successfully!"
        500:
          description: Internal Server Error
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  status: 
                    type: string
                    description: "CONTRACT_CONTEXTUAL_ERROR"
                  message: 
                    type: string
                    description: "Rental does not exist!"
components:
  schemas:
    Category:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        isDeleted:
          type: number
          format: int1
          default: 0
        isActive:
          type: number
          format: int1
          default: 1
        createBy:
          type: string
          default: "system"
        updateBy:
          type: string
        createAt:
          type: string
        updateAt:
          type: string
    Car:
      type: object
      properties:
        id:
          type: string
        category: 
          $ref: "#/components/schemas/Category"
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        pricePerDay:
          type: number
          format: double
        rentalStatus:
          type: number
          format: int1
        isDeleted:
          type: number
          format: int1
          default: 0
        isActive:
          type: number
          format: int1
          default: 1
        createBy:
          type: string
        updateBy:
          type: string
        createAt:
          type: string
        updateAt:
          type: string
    Customer:
      type: object
      properties:
        id:
          type: string
        firstName: 
          type: string
        lastName:
          type: string
        address:
          type: string
        mobilephone:
          type: string
        isDeleted:
          type: number
          format: int1
          default: 0
        isActive:
          type: number
          format: int1
          default: 1
        createBy:
          type: string
        updateBy:
          type: string
        createAt:
          type: string
        updateAt:
          type: string
    Rental:
      type: object
      properties:
        id:
          type: string
        car: 
          $ref: "#/components/schemas/Car"
        customer:
          $ref: "#/components/schemas/Customer"
        fromDate:
          type: string
        dueDate:
          type: string
        basePrice:
          type: number
          format: double
        discount:
          type: number
          format: double
          default: 0.00
        tax:
          type: number
          format: double
          default: 0.00
        totalPrice:
          type: number
          format: double
        status:
          type: number
          format: in1
          default: 0
        createBy:
          type: string
        updateBy:
          type: string
        createAt:
          type: string
        updateAt:
          type: string
...